Adds support for reading VCC files that do not have a timezone offset.
For example, a VCC file that has dateTime elements like this:
dateTime="2015-12-29T11:51:12" will be successfully read. Previously,
the dateTime element had to have a timezone offset, even if it was
zero: "2015-12-29T11:51:12+00:00”.

This was causing a problem because VCC files saved in the GMT timezone
do not have an offset.